requirements.txt
numpy==1.23.1
Pillow==9.2.0
torch==1.12.0
typing_extensions==4.3.0
